added missing test
